# apps/api/src/LiteNova.LitePress.Application.Read.Contracts/Posts/GetAllPosts/PostSummaryResult.cs
using LiteNova.LitePress.Application.Read.Contracts.Shared;

namespace LiteNova.LitePress.Application.Read.Contracts.Posts.GetAllPosts;

public sealed record PostSummaryResult(
    Guid PostId,
    string Title,
    string Slug,
    string? Excerpt,
    string? CoverImageUrl,
    string AuthorDisplayName,
    string PostState,
    DateTimeOffset CreatedAt,
    DateTimeOffset? PublishedAt,
    IReadOnlyList<TagSummaryResult> Tags
);